fix: keep focused client when readding clients from off output
Fixes #1133.

When a output is turned off, the clients go to another active
output. When this is the last output, and the output is then
reactivated in the future, the focus will be lost, because of the
order the clients are iterated in. The focused client is first in the
list, while the unfocused is last. This ensures the previously focused
client is enumerated as last, making it the currently focused client.